requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
aiofiles==23.2.1
Pillow==10.2.0
python-docx==1.1.0
openpyxl==3.1.2
PyPDF2==3.0.1
paho-mqtt==1.6.1
jinja2==3.1.3
